Implement FEBlendNeonApplier https://bugs.webkit.org/show_bug.cgi?id=243415
Reviewed by Simon Fraser.

This will return the FEBlend code back for ARMv7 Neon targets to its state
before r286152.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/cpu/arm/filters/FEBlendNeonApplier.cpp: Renamed from Source/WebCore/platform/graphics/cpu/arm/filters/FEBlendNEON.h.
(WebCore::FEBlendUtilitiesNEON::div255):
(WebCore::FEBlendUtilitiesNEON::normal):
(WebCore::FEBlendUtilitiesNEON::multiply):
(WebCore::FEBlendUtilitiesNEON::screen):
(WebCore::FEBlendUtilitiesNEON::darken):
(WebCore::FEBlendUtilitiesNEON::lighten):
(WebCore::FEBlendNeonApplier::applyPlatform const):
(WebCore::FEBlendNeonApplier::apply const):
* Source/WebCore/platform/graphics/cpu/arm/filters/FEBlendNeonApplier.h: Copied from Source/WebCore/platform/graphics/cpu/arm/filters/NEONHelpers.h.
* Source/WebCore/platform/graphics/cpu/arm/filters/NEONHelpers.h:
* Source/WebCore/platform/graphics/filters/FEBlend.cpp:
(WebCore::FEBlend::createSoftwareApplier const):
* Source/WebCore/platform/graphics/filters/FEBlend.h:

Canonical link: https://commits.webkit.org/253026@main
