+++ Source/WebCore/platform/graphics/cpu/arm/filters/NEONHelpers.h
@@ -24,8 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef NEONHelpers_h
-#define NEONHelpers_h
+#pragma once
 
 #if HAVE(ARM_NEON_INTRINSICS)
 
@@ -51,5 +50,3 @@ inline void storeFloatAsRGBA8(float32x4_t data, uint32_t* destination)
 } // namespace WebCore
 
 #endif // HAVE(ARM_NEON_INTRINSICS)
-
-#endif // NEONHelpers_h
