+++ Source/WebCore/platform/graphics/filters/FEBlend.h
@@ -42,11 +42,8 @@ private:
     FEBlend(BlendMode);
 
     unsigned numberOfEffectInputs() const override { return 2; }
-    
-    std::unique_ptr<FilterEffectApplier> createSoftwareApplier() const override;
 
-    void platformApplyNEON(unsigned char* srcPixelArrayA, unsigned char* srcPixelArrayB, unsigned char* dstPixelArray,
-                           unsigned colorArrayLength);
+    std::unique_ptr<FilterEffectApplier> createSoftwareApplier() const override;
 
     WTF::TextStream& externalRepresentation(WTF::TextStream&, FilterRepresentation) const override;
